refactor: consolidate classification pipelines into a unified ClassificationPipeline, removing separate binary, multiclass, and multioutput modules; enhance metric calculations and add comprehensive unit tests for all classification scenarios
