+++ coco_pipe/ml/__init__.py
@@ -1,9 +1,8 @@
-from .config import DEFAULT_CV, CLASSIFICATION_METRICS, REGRESSION_METRICS
+from .config import DEFAULT_CV, CLASSIFICATION_METRICS, REGRESSION_METRICS, BINARY_MODELS, MULTICLASS_MODELS
 from .utils import get_cv_splitter
 from .base import BasePipeline
-from .binary_classification import BinaryClassificationPipeline
-from .multiclass_classification import MultiClassPipeline
+from .classification import BinaryClassificationPipeline, MultiClassClassificationPipeline, MultiOutputClassificationPipeline, ClassificationPipeline
 __all__ = [
-    "DEFAULT_CV", "CLASSIFICATION_METRICS", "REGRESSION_METRICS",
-    "get_cv_splitter", "BasePipeline", "BinaryClassificationPipeline", "MultiClassPipeline"
+    "DEFAULT_CV", "CLASSIFICATION_METRICS", "REGRESSION_METRICS", "BINARY_MODELS", "MULTICLASS_MODELS",
+    "get_cv_splitter", "BasePipeline", "BinaryClassificationPipeline", "MultiClassClassificationPipeline", "MultiOutputClassificationPipeline", "ClassificationPipeline"
 ]
